change double to float

+++ db/migrate/20180725085830_create_products.rb
@@ -6,8 +6,8 @@ class CreateProducts < ActiveRecord::Migration[5.1]
       t.integer :category_id
       t.integer :user_id
       t.integer :status
-      t.double :price
-      t.double :acreage
+      t.float :price
+      t.float :acreage
       t.datetime :expiration_date
       t.integer :type
